Refactor heartbeat
